requirements.txt
Flask==2.0.2
PyPDF2==1.26.0
python-docx==0.8.11
pandas==1.4.4